% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Graph the population estimates of a country.}
\usage{
CountryPopulation(country)
}
\arguments{
\item{\code{country}}{A UN member nation name (string)}
}
\value{
A ggplot of \code{country} showing population estimates between 1950
and 2020
}
\description{
A graph of the UN population estimates for a country is returned when
entering a country name (must be a member of the United Nations) into the
function. If a country that does not exist or is not a UN member nation is
input, then an error is returned.
}
