% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{Attendance data for each World Cup}
\format{
a data frame with 22 observations
\describe{
  \item{WorldCup}{Host nation and year of hosting}
  \item{Totalattendance}{Total number of attendees during that World Cup}
  \item{Matches}{Number of matches played}
  \item{Averageattendance}{Average number of attendees at each match}
}
}
\source{
\url{https://en.wikipedia.org/wiki/FIFA_World_Cup}
}
\usage{
World_Cup
}
\description{
Attendance data for each World Cup
}
\keyword{datasets}
