% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{Estimates for population between 1950 and 2020 for 235 UN member nations}
\format{
a data frame with 235 observations
\describe{
  \item{Country}{Member nation of the United Nations}
  \item{Years}{Range from 1950 to 2020.}
}
}
\source{
UN Population Division
}
\usage{
WorldPopulation
}
\description{
Estimates for population between 1950 and 2020 for 235 UN member nations
}
\keyword{datasets}
